vbox: Add the option machinereadable and details to snapshot showvminfo
Summary:
There is already an option `--machinereadable` and `--details` for the command
`vboxmanage showvminfo [vm_name|uuid]`.
But the command `vboxmanage snapshot [vm_name|uuid] showvminfo [snapshot_name|uuid]` doesn't have them.
So, I did had them to the command for uniformization and because I need it to extract the FS.

Test Plan: Use the new option of the command line